Zip File containing current Code for GUI
UPDATES TO CODE:
REST_OVRVW_AUTMN is now treated as a menu of menu's. When the add menu button is clicked it creates a new form which is the menu of the specified type (i.e. Lunch, Dinner, etc.). Within the form there are two buttons (add food item and exit). Add Food Item button adds a button dynamically to the dynamically created form's FlowLayoutPanel. Exit closes the current form.